comment list moved inside hearing as its a dependant component
