+++ __tests__/comment.js
@@ -1,5 +1,5 @@
 import {renderIntoDocument} from 'react-dom/test-utils';
-import Comment from '../src/components/Comment';
+import Comment from '../src/components/Hearing/Comment';
 import {findDOMNode} from 'react-dom';
 import {wireComponent} from '../test-utils';
