+++ src/components/Hearing/CommentList/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {injectIntl, FormattedMessage} from 'react-intl';
-import Comment from '../components/Hearing/Comment';
+import Comment from '../Comment';
 
 export class CommentList extends React.Component {
   render() {
